change delete to svg, add confirmation

+++ src/ui.js
@@ -1,6 +1,9 @@
 import { manager } from "./tasks";
 import { saveTasks } from "./storage";
 import { createProjectList } from "./sidebar";
+import deleteButton from "./images/x-square.svg";
+
+const deleteSVG = '<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="#89a9a9" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-x-square"><rect x="3" y="3" width="18" height="18" rx="2" ry="2"></rect><line x1="9" y1="9" x2="15" y2="15"></line><line x1="15" y1="9" x2="9" y2="15"></line></svg>';
 
 export function createForm() {
     const taskListContainer = document.querySelector("#taskListContainer");
@@ -106,12 +109,15 @@ function appendToList(task, taskIndex) {
         taskItem.checked = task.completed;
     }
 
-    const divDelete = document.createElement('button');
-    divDelete.textContent = "X";
-    divDelete.classList = 'taskDelete';
-    divDelete.addEventListener('click', () => {
-        manager.deleteTask(taskItem.getAttribute('taskIndex'));
-        refreshTaskList();
+
+    const imgDelete = document.createElement('img');
+    imgDelete.src = deleteButton;
+    imgDelete.classList = 'taskDelete';
+    imgDelete.addEventListener('click', () => {
+        if (confirm('Delete task? (THIS IS PERMANENT)')){
+            manager.deleteTask(taskItem.getAttribute('taskIndex'));
+            refreshTaskList();
+        }
     });
 
     taskItem.addEventListener('change', () => {
@@ -121,7 +127,7 @@ function appendToList(task, taskIndex) {
 
     taskDiv.appendChild(taskItem);
     taskDiv.appendChild(taskLabel);
-    taskDiv.appendChild(divDelete);
+    taskDiv.appendChild(imgDelete);
     if (task.completed){
         listComplete.appendChild(taskDiv);
     } else {
